requirements.txt
numpy~=1.22.3
matplotlib~=3.5.2
Pillow~=9.1.0
tensorflow~=2.8.0
sklearn~=0.0
scikit-learn~=1.0.2